requirements.txt
anthropic==0.18.1
pyyaml==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3 